chore(baseline-crop): add documentation and clean up the code
