+++ packages/layout/src/layout.utils.ts
@@ -1,75 +1,7 @@
-import { __DEV__, isArray, get as getFromTheme } from "@chakra-ui/utils"
+import { __DEV__, get as getFromTheme } from "@chakra-ui/utils"
+import { convertToArray } from "./baselineCrop/breakpoint.utils"
 
-type Breakpoints = { [key: string]: string }
-
-function convertBreakpointsIntoArray(breakpoints: Breakpoints): string[] {
-  return Object.keys(breakpoints)
-    .sort((a, b) => parseInt(breakpoints[a]) - parseInt(breakpoints[b]))
-    .map(key => breakpoints[key])
-}
-
-/**
- * convertToArray takes a given style property, like `lineHeight`, and converts it into an
- * equivalent styled-system responsive array
- *
- * @param theme - the styled-system theme object
- * @param value - the style value to convert to an array
- *
- * @example Examples can be found in the associated test file
- */
-export function convertToArray<T>(
-  theme: object,
-  value: T | T[] | { [key: string]: T },
-): T[] {
-  const breakpoints: Breakpoints = getFromTheme(theme, "breakpoints")
-
-  if (isArray(value)) {
-    return value
-  } else if (value === undefined) {
-    return []
-  } else if (typeof value === "object") {
-    if (__DEV__) {
-      // Perform some extra checks in development mode
-      if (isArray(breakpoints)) {
-        throw new Error(
-          "When using `isBaselineCropped`, you must define your breakpoints as an object.",
-        )
-      }
-      const breakpointUnits = Object.keys(breakpoints).map(key =>
-        breakpoints[key].replace(/^\d+/, ""),
-      )
-      if (!breakpointUnits.every(unit => unit === breakpointUnits[0])) {
-        throw new Error(
-          `
-          When using isBaselineCropped, you must define every breakpoint using the same unit.
-
-          You provided the following breakpoints: ${JSON.stringify(breakpoints)}
-          We calculated these units "${breakpointUnits}" which are not all the same.
-          `,
-        )
-      }
-    }
-    // In order for the text crop feature to work, we must know the _order_ of the breakpoints,
-    // so we convert the object breakpoints into
-    const breakpointArray = convertBreakpointsIntoArray(breakpoints)
-
-    /* If we receive a styled-system responsive style object, we want to manually convert it into
-     an equivalent styled-system responsive style array */
-    const breakpointKeys = Object.keys(value)
-    const breakpointIndices = breakpointKeys.map(breakpoint =>
-      breakpointArray.indexOf(breakpoints[breakpoint]),
-    )
-    const highestBreakpointIndex = Math.max(...breakpointIndices)
-    const result: T[] = new Array(highestBreakpointIndex + 1).fill(null)
-    breakpointKeys.forEach(breakpoint => {
-      const i = breakpointArray.indexOf(breakpoints[breakpoint])
-      result[i] = (value as { [key: string]: T })[breakpoint]
-    })
-    return result
-  } else {
-    return [value]
-  }
-}
+export type Breakpoints = { [key: string]: string }
 
 export type LineHeight =
   | number
@@ -84,14 +16,19 @@ export type FontFamily =
   | { [key: string]: string }
   | undefined
 
+// If baselineCrop is enabled, and the user does not explicitly pass a line height or font family, we will use these values
 const DEFAULT_LINE_HEIGHT = 1.2
 const DEFAULT_FONT_FAMILY = "body"
+
+// Used for measuring the baseline crop settings
+const DEFAULT_CROP_FONT_SIZE = 48
+const DEFAULT_CROP_LINE_HEIGHT = 1.2
 
 function preprocessLineHeight(
   theme: object,
   rawLineHeight: number | string | null | undefined = DEFAULT_LINE_HEIGHT,
 ): number | null {
-  const invalidLineHeight = () => {
+  const handleInvalidLineHeight = () => {
     if (__DEV__) {
       throw new Error(
         `When using the "isBaselineCropped" prop, you must only use numeric values for line-height or the string "normal".`,
@@ -101,7 +38,9 @@ function preprocessLineHeight(
 
   if (typeof rawLineHeight === "undefined") {
     return DEFAULT_LINE_HEIGHT
-  } else if (rawLineHeight === null) {
+  }
+
+  if (rawLineHeight === null) {
     return null
   }
 
@@ -112,7 +51,7 @@ function preprocessLineHeight(
       : rawLineHeight
 
   if (typeof lineHeight === "undefined") {
-    invalidLineHeight()
+    handleInvalidLineHeight()
   }
 
   /*
@@ -135,7 +74,7 @@ function preprocessLineHeight(
     */
 
     if (isNaN(parsed)) {
-      invalidLineHeight()
+      handleInvalidLineHeight()
       return DEFAULT_LINE_HEIGHT
     }
 
@@ -150,26 +89,37 @@ interface TextCropValues {
   bottomCrop: number
 }
 
+function getBaselineCropSettings(theme: object, fontFamily: string) {
+  const baselineCropSettings = getFromTheme(theme, `baselineCrop.${fontFamily}`)
+
+  if (!baselineCropSettings) {
+    if (__DEV__) {
+      throw new Error(`
+      In order to use the baselineCrop prop, you must manually add baselineCrop measurements into your theme file.
+      See the Text docs on the Chakra-ui website for instructions on how to measure a font: https://chakra-ui.com/text
+      `)
+    }
+  }
+
+  return {
+    ...baselineCropSettings,
+    cropFontSize: DEFAULT_CROP_FONT_SIZE,
+    cropLineHeight: DEFAULT_CROP_LINE_HEIGHT,
+  }
+}
+
 function calculateTextCrop(
   theme: object,
   { lineHeight, fontFamily }: { lineHeight: number | null; fontFamily: string },
 ): TextCropValues | null {
   if (lineHeight === null) return null
 
-  const baselineCropSettings = getFromTheme(theme, `baselineCrop.${fontFamily}`)
-
-  if (!baselineCropSettings) {
-    // TODO link directly to the page where we can perform those measurements
-    if (__DEV__) {
-      throw new Error(`
-      In order to use the baselineCrop prop, you must manually add baselineCrop settings into your theme file.
-      See the Text docs on the Chakra-ui website for instructions on how to measure these settings.
-      `)
-    }
-  }
-  const { topCrop, bottomCrop } = baselineCropSettings
-  const cropFontSize = 41
-  const cropLineHeight = 1.2
+  const {
+    topCrop,
+    bottomCrop,
+    cropFontSize,
+    cropLineHeight,
+  } = getBaselineCropSettings(theme, fontFamily)
 
   const dynamicTopCrop: number =
     Math.max(topCrop + (lineHeight - cropLineHeight) * (cropFontSize / 2), 0) /
@@ -186,7 +136,18 @@ function calculateTextCrop(
   }
 }
 
-export function fillArray<T>(values: T[], length: number): T[] {
+/**
+ * Takes a responsive style array and fills in implicit nulls with explicit values.
+ *
+ * @param values - array of any values, may contain nulls
+ * @param length - the desired length of the resulting array
+ *
+ * @example fillResponsiveArray([1, null]) // [1, 1]
+ * @example fillResponsiveArray([1, null, 2]) // [1, 1, 2]
+ * @example fillResponsiveArray([null, 2]) // [null, 2]
+ * @example fillResponsiveArray([1, null, null, 2, null]) // [1, 1, 1, 2, 2]
+ */
+export function fillResponsiveArray<T>(values: T[], length: number): T[] {
   const result = []
   let mostRecentValue = values[0]
 
@@ -234,36 +195,32 @@ export function calculateBaselineCrop(
     fontFamily?: FontFamily
   },
 ) {
-  // Transform single values into singleton arrays
-  const unprocessedLineHeights = convertToArray(
+  // Normalize lineHeight and fontFamily props into comparable structures
+  let lineHeightResponsiveArray = convertToArray(
     theme,
     props.lineHeight ?? DEFAULT_LINE_HEIGHT,
-  )
-  const unprocessedFontFamilies = convertToArray(theme, props.fontFamily)
-
-  // Turn all line heights into numbers (look up string values from the theme)
-  let processedLineHeights = unprocessedLineHeights.map(rawLineHeight =>
-    preprocessLineHeight(theme, rawLineHeight),
-  )
+  ).map(rawLineHeight => preprocessLineHeight(theme, rawLineHeight))
+  const fontFamilyResponsiveArray = convertToArray(theme, props.fontFamily)
 
   const longestStyleArray = Math.max(
-    unprocessedFontFamilies.length,
-    processedLineHeights.length,
+    fontFamilyResponsiveArray.length,
+    lineHeightResponsiveArray.length,
   )
 
   // fill in all nulls into the font-family array such that we have the same number of line-heights and font-families
-  const processedFontFamilies = fillArray(
-    unprocessedFontFamilies,
+  const processedFontFamilies = fillResponsiveArray(
+    fontFamilyResponsiveArray,
     longestStyleArray,
   ).map(family => family ?? DEFAULT_FONT_FAMILY)
 
   // fill in all nulls into the line-height array such that we have the same number of line-heights and font-families
-  processedLineHeights = fillArray(processedLineHeights, longestStyleArray).map(
-    lineHeight => lineHeight ?? DEFAULT_LINE_HEIGHT,
-  )
+  lineHeightResponsiveArray = fillResponsiveArray(
+    lineHeightResponsiveArray,
+    longestStyleArray,
+  ).map(lineHeight => lineHeight ?? DEFAULT_LINE_HEIGHT)
 
   // Calculate the number of pixels needed to crop from the top/bottom of the text
-  const textCropValues = processedLineHeights.map((lineHeight, index) => {
+  const textCropValues = lineHeightResponsiveArray.map((lineHeight, index) => {
     const fontFamily =
       processedFontFamilies[Math.min(processedFontFamilies.length - 1, index)]
 
@@ -274,7 +231,7 @@ export function calculateBaselineCrop(
   })
 
   return {
-    lineHeight: processedLineHeights,
+    lineHeight: lineHeightResponsiveArray,
     transform: textCropValues.map(values =>
       values ? `translateY(${values.bottomCrop.toFixed(5)}em)` : null,
     ),
