+++ packages/layout/src/tests/layout.utils.test.ts
@@ -1,11 +1,7 @@
-import {
-  convertToArray,
-  fillArray,
-  calculateBaselineCrop,
-} from "./layout.utils"
+import { fillResponsiveArray, calculateBaselineCrop } from "../layout.utils"
 import baseTheme from "@chakra-ui/preset-base"
 
-describe("fillArray()", () => {
+describe("fillResponsiveArray()", () => {
   const cases = [
     [[], []],
     [["a"], ["a"]],
@@ -29,41 +25,19 @@ describe("fillArray()", () => {
 
   cases.forEach(([input, output], index) => {
     it(`should work for case ${index}`, () => {
-      const result = fillArray(input, input.length)
+      const result = fillResponsiveArray(input, input.length)
       expect(result).toEqual(output)
     })
   })
 
   it("should work for longer length", () => {
-    const result = fillArray(["a", null, "b"], 5)
+    const result = fillResponsiveArray(["a", null, "b"], 5)
     expect(result).toEqual(["a", "a", "b", "b", "b"])
   })
 
   it("should work for empty array", () => {
-    const result = fillArray([], 3)
+    const result = fillResponsiveArray([], 3)
     expect(result).toEqual([null, null, null])
-  })
-})
-
-describe("convertToArray()", () => {
-  const theme: any = {
-    breakpoints: { sm: "2em", md: "5em", lg: "10em" },
-  }
-
-  it("should work for a single value", () => {
-    expect(convertToArray(theme, 1.3)).toEqual([1.3])
-    expect(convertToArray(theme, "34rem")).toEqual(["34rem"])
-  })
-  it("should work for an array", () => {
-    expect(convertToArray(theme, [1.2, "3rem"])).toEqual([1.2, "3rem"])
-    expect(convertToArray(theme, [null, "3rem"])).toEqual([null, "3rem"])
-  })
-  it("should work for an object", () => {
-    expect(convertToArray(theme, { sm: 1 })).toEqual([1])
-    expect(convertToArray(theme, { lg: 1 })).toEqual([null, null, 1])
-    expect(convertToArray(theme, { sm: 1, md: 2 })).toEqual([1, 2])
-    expect(convertToArray(theme, { sm: 1, lg: 2 })).toEqual([1, null, 2])
-    expect(convertToArray(theme, { md: 1, lg: 2 })).toEqual([null, 1, 2])
   })
 })
