+++ website/pages/heading.mdx
@@ -8,7 +8,7 @@ styles as well. It renders an `<h2>` tag by default.
 ## Import
 
 ```js
-import { Heading } from "@chakra-ui/core";
+import { Heading } from "@chakra-ui/core"
 ```
 
 ## Usage
@@ -83,6 +83,12 @@ prop. No need to write `css` or `styled()`
   </Button>
 </Box>
 ```
+
+## Baseline Crop
+
+The `Heading` (and `Text`) components support a prop called `baselineCrop`. This
+prop allows us to remove the extra white space that normally appears above and
+below text. See the [documentation for the `Text` component](/text) for details.
 
 ## Props
